PR fixes courtsey of Eric
