+++ src/commands/createNewProject/IProjectCreator.ts
@@ -11,6 +11,7 @@ import { promptForProjectRuntime } from "../../ProjectSettings";
 export abstract class ProjectCreatorBase {
     public deploySubpath: string = '';
     public preDeployTask: string = '';
+    public excludedFiles: string | string[] = '';
     public abstract templateFilter: TemplateFilter;
 
     protected readonly functionAppPath: string;
