+++ src/commands/createNewProject/JavaScriptProjectCreator.ts
@@ -16,6 +16,9 @@ export const funcNodeDebugEnvVar: string = 'languageWorkers:node:arguments';
 export class JavaScriptProjectCreator extends ScriptProjectCreatorBase {
     public readonly templateFilter: TemplateFilter = TemplateFilter.Verified;
     public readonly deploySubpath: string = '.';
+    // "func extensions install" task creates C# build artifacts that should be hidden
+    // See issue: https://github.com/Microsoft/vscode-azurefunctions/pull/699
+    public readonly excludedFiles: string | string[] = ['obj', 'bin'];
 
     public readonly functionsWorkerRuntime: string | undefined = 'node';
