+++ src/constants.ts
@@ -12,6 +12,7 @@ export const templateFilterSetting: string = 'templateFilter';
 export const deploySubpathSetting: string = 'deploySubpath';
 export const templateVersionSetting: string = 'templateVersion';
 export const preDeployTaskSetting: string = 'preDeployTask';
+export const filesExcludeSetting: string = 'files.exclude';
 
 export enum ProjectLanguage {
     Bash = 'Bash',
